Init
There is a basic dice game that relies on random numbers generated by drand. One-tenth of the game's earnings will go into the game's bonus account until someone with the game capability withdraws it into the bonus pool and waits for distribution. During distribution, Tokens need to be staked in the bonus pool until the payout time arrives to claim both the Tokens and the distribution.